Fix file:// handling under Windows
Windows paths do not start with a slash, so we add an extra slash to
separate the host from the path in file:// urls. Otherwise "D:" is
parsed as the host segment in the URI.

The path for those URLs now starts with "/", so we ignore that leading
character when using the URI's path.

This reduces Windows CI spec failures from 429 to 355.